feat(fonts): add NotesEsa font to markers
