+++ src/scripts/custom.d.ts
@@ -7,3 +7,8 @@ declare module '*.svg' {
   const content: string;
   export default content;
 }
+
+declare module '*.otf' {
+  const content: string;
+  export default content;
+}
